+++ pg_documentdb/src/configs/feature_flag_configs.c
@@ -80,6 +80,9 @@ bool DefaultUseCompositeOpClass = DEFAULT_USE_NEW_COMPOSITE_INDEX_OPCLASS;
 #define DEFAULT_ENABLE_INDEX_ORDERBY_PUSHDOWN false
 bool EnableIndexOrderbyPushdown = DEFAULT_ENABLE_INDEX_ORDERBY_PUSHDOWN;
 
+#define DEFAULT_ENABLE_INDEX_ORDERBY_PUSHDOWN_LEGACY false
+bool EnableIndexOrderbyPushdownLegacy = DEFAULT_ENABLE_INDEX_ORDERBY_PUSHDOWN_LEGACY;
+
 #define DEFAULT_ENABLE_DESCENDING_COMPOSITE_INDEX true
 bool EnableDescendingCompositeIndex = DEFAULT_ENABLE_DESCENDING_COMPOSITE_INDEX;
 
@@ -197,6 +200,9 @@ bool EnableIndexOnlyScan = DEFAULT_ENABLE_INDEX_ONLY_SCAN;
 
 #define DEFAULT_ENABLE_RANGE_OPTIMIZATION_COMPOSITE false
 bool EnableRangeOptimizationForComposite = DEFAULT_ENABLE_RANGE_OPTIMIZATION_COMPOSITE;
+
+#define DEFAULT_USE_PG_STATS_LIVE_TUPLES_FOR_COUNT true
+bool UsePgStatsLiveTuplesForCount = DEFAULT_USE_PG_STATS_LIVE_TUPLES_FOR_COUNT;
 
 /* FEATURE FLAGS END */
 
@@ -491,6 +497,14 @@ InitializeFeatureFlagConfigurations(const char *prefix, const char *newGucPrefix
 		PGC_USERSET, 0, NULL, NULL, NULL);
 
 	DefineCustomBoolVariable(
+		psprintf("%s.enableIndexOrderbyPushdownLegacy", newGucPrefix),
+		gettext_noop(
+			"Whether to enable the prior index sort on the new experimental composite index opclass"),
+		NULL, &EnableIndexOrderbyPushdownLegacy,
+		DEFAULT_ENABLE_INDEX_ORDERBY_PUSHDOWN_LEGACY,
+		PGC_USERSET, 0, NULL, NULL, NULL);
+
+	DefineCustomBoolVariable(
 		psprintf("%s.enableDescendingCompositeIndex", newGucPrefix),
 		gettext_noop(
 			"Whether to enable descending composite index support"),
@@ -553,5 +567,13 @@ InitializeFeatureFlagConfigurations(const char *prefix, const char *newGucPrefix
 			"Whether to enable range optimization for composite indexes."),
 		NULL, &EnableRangeOptimizationForComposite,
 		DEFAULT_ENABLE_RANGE_OPTIMIZATION_COMPOSITE,
+		PGC_USERSET, 0, NULL, NULL, NULL);
+
+	DefineCustomBoolVariable(
+		psprintf("%s.usePgStatsLiveTuplesForCount", newGucPrefix),
+		gettext_noop(
+			"Whether to use pg_stat_all_tables live tuples for count in collStats."),
+		NULL, &UsePgStatsLiveTuplesForCount,
+		DEFAULT_USE_PG_STATS_LIVE_TUPLES_FOR_COUNT,
 		PGC_USERSET, 0, NULL, NULL, NULL);
 }