+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### documentdb v1.107-0 (Unreleased) ###
 - Support sort by _id against the _id index using the enableIndexOrderbyPushdown flag *[Feature]*.
+- Improvements to explain for various scan types *[Feature]*.
 
 ### DocumentDB v0.106.0-ferretdb-2.5.0 (August 12, 2025) ###